+++ app/handlers/apiv1/tag.go
@@ -2,6 +2,7 @@ package apiv1
 
 import (
 	"github.com/getfider/fider/app/actions"
+	"github.com/getfider/fider/app/models"
 	"github.com/getfider/fider/app/pkg/web"
 )
 
@@ -31,6 +32,50 @@ func UnassignTag() web.HandlerFunc {
 		}
 
 		err := c.Services().Tags.UnassignTag(input.Tag, input.Post)
+		if err != nil {
+			return c.Failure(err)
+		}
+
+		return c.Ok(web.Map{})
+	}
+}
+
+// CreateEditTag creates a new tag on current tenant
+func CreateEditTag() web.HandlerFunc {
+	return func(c web.Context) error {
+		input := new(actions.CreateEditTag)
+		if result := c.BindTo(input); !result.Ok {
+			return c.HandleValidation(result)
+		}
+
+		var (
+			tag *models.Tag
+			err error
+		)
+
+		if input.Model.Slug != "" {
+			tag, err = c.Services().Tags.Update(input.Tag, input.Model.Name, input.Model.Color, input.Model.IsPublic)
+		} else {
+			tag, err = c.Services().Tags.Add(input.Model.Name, input.Model.Color, input.Model.IsPublic)
+		}
+
+		if err != nil {
+			return c.Failure(err)
+		}
+
+		return c.Ok(tag)
+	}
+}
+
+// DeleteTag deletes anexisting tag
+func DeleteTag() web.HandlerFunc {
+	return func(c web.Context) error {
+		input := new(actions.DeleteTag)
+		if result := c.BindTo(input); !result.Ok {
+			return c.HandleValidation(result)
+		}
+
+		err := c.Services().Tags.Delete(input.Tag)
 		if err != nil {
 			return c.Failure(err)
 		}
