+++ public/components/ShowTag.scss
@@ -10,6 +10,10 @@
   border: 0;
   box-shadow: inset 0 -1px 0 rgba(27,31,35,0.12);
 
+  i.icon {
+    font-size: $font-size-micro;
+  }
+
   &:last-child {
     margin-top: 0;
     margin-bottom: 0;
